Correctly destoy RandomGenerator singleton
Change the implementation of Couchbase::RandomGenerator to be
correctly destoyed on process termination, removing the previous
memory leak (still reachable) which existed.

Note that C++11 guarantees that concurrent execution of the
initialization of the static member is thread-safe, so we can
significantly simplify the implementation by just using a static
member in the constructor.

Change-Id: Ifc169112d0d1fc870319df6bffb4eb301c38594f
Reviewed-on: http://review.couchbase.org/52778
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>