MB-16604 [BP 3.x]: Fix JSON_checker leak upon invalid UTF-8
A memory leak has been identified in the checkUTF8JSON function which
is used to determine the datatype field for document
mutations. Specifically, the leak occurs when validating UTF8 contents
of a JSON document. If an invalid UTF8 escape sequence is encountered
then we correctly mark the document as "binary" (non-JSON), however we
leak 52 bytes.

This is due to an error in the JSON_checker cleanup code - it will
free that memory if the actual JSON structure was invalid, but not if
it only encountered a bad UTF-8 value.

In general any document which:
1. Contains a byte of '"' (i.e. ASCII 0x22)
2. Followed (not necessary immediately) by 0xf8...0xff
3. And hasn't already failed JSON validation

could trigger this issue.

Change-Id: I6ff54242a7427f520891446cb597a9e3dd5188e1
Reviewed-on: http://review.couchbase.org/56385
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>