MB-16984: Move ep-engine Histogram to platform
ep-engine has a histo.h file which contains code to measure the
duration of code blocks, and to record that information into a
histogram. This functionality is required inside memcached to support
sub-document API parse timings. Move / refactor the code form
ep-engine into platform.

Note the code required some small changes to live in platform -
previously it used ep-engine's LOG() function directly, which isn't
available in platform. The call to LOG() has been replaced with a call
to fprint(stderr, ...) - given this is (currently) debug-only
functionality this seems acceptable.

Change-Id: Idcb02d40167b2854422afdbdb80b13c3c4e266eb
Reviewed-on: http://review.couchbase.org/57509
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>