Refactor: Remove plugable allocation API (not used)
Use a fixed one that use malloc/calloc/strdup/free but throw
an exception if allocation fails (not all of the existing paths
checked the return code which would lead to nullpointer
exceptions).

In addition to that not all methods had an API which would allow
to propagate the error back up.

Change-Id: I363addcd22a227acdad4a1059ab4cac67786c70c
Reviewed-on: http://review.couchbase.org/63024
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>