MB-20769: Use cbmalloc for cJSON
cJSON performs memory allocation internally, and also allocates memory
which is returned (either directly or indirectly) to clients. For both
cases it uses a set of malloc functions to perform allocation, which
default to the standard system malloc/free etc.

This patch changes the default malloc functions to be cbmalloc (the
user can still use the hook API to overide to something different). It
also fixes a couple of instances where the system functions were
incorrectly hard-coded.

Change-Id: Ibc81ee0342fdd864f653f3a816463c9c16811242
Reviewed-on: http://review.couchbase.org/67815
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>