MB-20834: Add Couchbase::get_available_cpu_count
Add a new helper function to platform to return the number of logical
threads (CPUs) this process has access to - i.e. the maximum number of
concurrent threads of execution available.

This is essentially the code from memcached's
get_number_of_worker_threads(), enhanced to use sched_getaffinity() if
available. This is particularly important on Docker which uses cgroups
to control the number of CPUs the container has access to, and by not
using that, we create "too many" worker threads.

Change-Id: I3e3b91eecc51aea298ae9aceb9e8c6d3f16b7612
Reviewed-on: http://review.couchbase.org/67490
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>