Add helper functions from memcached
memcached_testapp contains an implementation to determine the
current working directory. We've had the use for the method
elsewhere so it should live in dirutils.

The C API for cb_mktemp is annoying to use from a C++ context
as the callers need to add to the patternmask for it to
generate the uniqueness. memcached_testapp had a wrapper
function for this to make it easier to use from C++.

Change-Id: I0744a0ee09807d133915a3ba7bfe1db05e9f3d6e
Reviewed-on: http://review.couchbase.org/68386
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>