Add cb::extmeta with methods to build and parse the extmeta
This code is intended to replace the encoding in ep_engine, but
also allow memcached (and others) to build (and parse) the
metadata.

The intended use case for this is to extend it to add metadata
type for:

   * xattr size
   * collection length

The Builder class may be used to construct the meta
stream, and the Parser is used to decode the meta stream
into the different values.

Change-Id: Ia4c08618370cdabb2ff4944ee29ebf104f22612a
Reviewed-on: http://review.couchbase.org/68915
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>