MB-21436: Add a guard to cb_join_thread from waiting for itself
We've seen problems where we've had global std::unique_ptr's which
had to run destructors which waited for threads be run on a "random"
thread causing a deadlock (the actual use was in memcached with the
parent monitor thread). The C++ runtime just picked a random thread
to run these destructors, and sometimes the parent monitor thread
would run them. We've refactored the code in memcached so that object
is no longer global, but to avoid facing the problem at a later time
we should add a guard here. (and it doesn't make any sense from
a logical perspective to wait for the current thread to be done ;-)

Change-Id: Ieef1ce83f53fabd5a77c0df1499641b094e23eb4
Reviewed-on: http://review.couchbase.org/69034
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>