MB-20079: Provide ability to use std::chrono::steady_clock
Currently we only have the ability to use gethrtime(void) which uses
clock_gettime(CLOCK_REALTIME_FAST ...)

This returns the current view of wall-clock time and therefore can
jump forwards and backwards as the system wall clock is changed.

The steady_clock is not related to wall clock time and cannot be
decreased.  This is useful when we want a task to run at a set
interval regardless if the system wall clock is changed.

Unfortunately there is an issue with the accuracy of steady_clock on
Windows when built with VS 2013 or earlier.  Therefore an abstraction
of ProcessClock is provided, which on non-Windows or when compiling
with 2015+ maps directly onto std::chrono::steady_clock.  However if
compiling on Windows using VS 2013 or below ProcessClock maps to our
own implementation of steady_clock.

Change-Id: I2fc9688abb782fe2c9e80efb6da840be3643d4a5
Reviewed-on: http://review.couchbase.org/69844
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>