Copy make_unique() into std namespace
When the backport of C++14 make_unique() was added to platform, it
incorrectly was added to the global namespace instead of std. This
means it isn't a drop-in replacement - it /should/ be in the std
namespace.

However to fix this we need to change our existing code to call
std::make_unique() instead of make_unique at the same time as moving
the definition here. Given the uses are in different repos to platform
(ep-engine), this is tricky.

Therefore, add an additional copy of make_unique (in the correct
namespace) in this first patch. Callers can then be updated to use
std::make_unique(), and then finally we remove the temporary copy.

Change-Id: I2909bb47966e979f2c25aa0d9199688b7e0a91b7
Reviewed-on: http://review.couchbase.org/70683
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>