MB-22009: Remove asprintf and vasprintf
asprintf uses C-style allocation so relies on the
user to free the allocated memory with free() rather
than cb_free(). The C++ std::string is now preferred
so there is no need for asprintf.

Remove the definitions of asprintf and vasprintf from
win32/strings.h.

Delete strings.c as the only implementations were asprintf
and vasprintf.

Remove the tests for asprintf and vasprintf

Change-Id: Ia8ee78ff7f2623888643f1622b00d5898b186891
Reviewed-on: http://review.couchbase.org/71853
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>