MB-22181: Refactor Histogram to use unique_ptrs
Refactors Histogram to use unique pointers and other more modern C++
features and style.

Replaces Histogram<T>::iterator as it was mostly boilerplate.
Replaced with usings.

Changes bins to be a vector of unique pointers rather than raw pointers.
This is to enable a simpler move constructor and to modernise the
class.

Change-Id: I75cdefc6d988e9f5c9f8b11345748e4959a8af58
Reviewed-on: http://review.couchbase.org/72113
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>