NonNegativeCounter: Allow underflow policy to be customized
Add an additional template parameter to NonNegativeCounter named
UnderflowPolicy which can be used to customize the behaviour of this
class when it underflows.

The policy defaults to ClampAtZeroUnderflowPolicy (which is the
previous behavour), but it also adds ThrowExceptionUnderflowPolicy for
use in debugging / situations where we want to know if the counter
would have underflown.

Change-Id: I576c250a1627203b386daba23511b11555a49684
Reviewed-on: http://review.couchbase.org/76636
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>