Create cbsockets (cb::net) containing socket wrappers
This is a wrapping layer on top of BSD sockets to allow for logging all
of the data being transferred. It was initially created in order to
track down why one of the pipes in moxi seemed to get out of sync
protocol wise, but may be useful in the future. This extra layer cause
an extra lookup of an atomic variable to determine if logging should
be performed or not which shouldn't affect the performance too much.

As it was initially created for locating a bug in moxi, a C API exists
with a 'cb_' prefix. It all wraps into methods in cb::net namespace.

Each method works as documented in the in the OS.

@todo open up for a callback to be called to allow the app to
determine if a particular socket should be logged or not (only if
the global logger flag is set)

@todo As each application may open and close a lot of sockets during
its lifetime, the current version _deletes_ the files when the application
close the socket. The use case when I developed this was that the program
would crash and generate a coredump (but it could open/send-receive/close
a lot of data which could generate tens of GB of data first).

One could use tcpdump to collect the same info, but I failed to get
tcpdump create a pcap file I know how to parse when using the loopback
interface. Whipping up this seemed a lot easier ;-)

Change-Id: I9f68912c61fdc8d20998bf3e6a9e147b7d77063b
Reviewed-on: http://review.couchbase.org/78518
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>