Add ProcessClockSource for dependency injection of sources of ProcessClock
This patch adds a pure interface `ProcessClockSource` with a single
method `now()` for getting the current time as a ProcessClock::time_point.

The patch also adds a default implementation that maps to ProcessClock::now()
called `DefaultProcessClockSource` and a default instance of it that can
be retreived via `defaultProcessClockSource()`.

Simultaneously this patch cleans up the ProcessClock API by placing
it within the `cb` namespace and adding `using` declarations for previously
global and used names.

Change-Id: Ie2c164ac34fe7d66b8210b5c5188837788b775d6
Reviewed-on: http://review.couchbase.org/79021
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>