Add user-defined literal for const_char_buffer
C++11 introduces user defined literals and Windows finally
supports them.

This patch adds the literal operator ""_ccb for creating
objects of cb::const_char_buffer from string literals.

Example:

    namespace Collections {
        const auto SystemEventPrefix = "$collections"_ccb;
    }

The literal is prefixed with '_' as required by the language
specification. The literal is not placed in a namespace like the
standard library literals since we are not likely to encounter
naming conflicts within KV-Engine and it makes it easier to use
the literal in headers (where we typically ban `using namespace`
declarations).

Change-Id: I3048deb9858e280ccdd88f76fa89bf5038c38662
Reviewed-on: http://review.couchbase.org/79070
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>