MB-24634: Add cb::Pipe
The Pipe class is a buffered pipe where you may insert data in one
end, and read it back out from the other (like a normal pipe).

Instead of using customized iosteram buffers and allows for that kind
of use, this class provides a "produce" method and a "consume" method
where you provide a callback which receives the memory area it may operate
on so that it can perform more optimal IO by populating the buffer
directly:

   pipe.produce([&sock](void* data, size_t size) -> ssize_t {
       return cb::net::recv(sock, data, size, 0);
   });

   And now you may use the data by calling:

   pipe.consume([](const void* data, size_t size) -> ssize_t {
       // do whatever we want with the data
       return nbytes;
   });

The return value from the consume method is the number of bytes you've
consumed (so that they may be dropped from the pipe). It means that if
you just want to look at the bytes you would return 0, and they would
still be present the next time you invoke the consume method.

Implementation details:

The Pipe is implemented by using allocating an underlying buffer of a
fixed size. The caller may grow this buffer by calling ensureCapacity,
which would:

   * return immediately if available segment at the end is big enough
   * pack the buffer if the free segment at the end and the beginning
     is big enough
   * reallocate the underlying buffer if we need more space

As you can see from the above the last two bullets would invalidate
all addresses previously provided to the "produce" and "consume" methods
which means that you _cannot_ keep pointers into this buffer and use
ensureCapacity.

Data written to the pipe is always put at the end of the allocated
buffer, and the write_head is moved forward. Data is always read from
the read_head, and once consumed the read head is moved forward. Whenever
the read_head catch up with the write_head they're both set to 0 (the
beginning of the buffer).

Change-Id: Ibe4f65bc5135ff3765a051ed013c59a330dffefe
Reviewed-on: http://review.couchbase.org/79088
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>