Add tighter integration with valgrind for cb::Pipe
The underlying backing store for the pipe is allocated in one
continuous chunk, and by default valgrind detects any buffer
over/underflow on that buffer.

We do however have strict requirements on where we're allowed
to read/write in the buffer, and valgrind allows for a tighter
integration by us telling it if the memory should be readable
or not.

This patch makes sure that when you try to write data into
the buffer you may _only_ access the portion of the buffer
where new data may be inserted. At any other time, the client
may access the portion of the data previously written into
the buffer.

Change-Id: Id22b93d6b229938816fbe1e06bdc757a0d6d7992
Reviewed-on: http://review.couchbase.org/79475
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Will Gardner <willg@rdner.io>