Add cb::make_array
This introduces cb::make_array, included from the boost::make_array
(which is very similar to the std::make_array in the C++ Library
Fundamentals TS V2).

The purpose of cb::make_array is to make declaration of compile time
arrays easier by allowing inference of both the common type of its
members and the size of the array.

Example usage:

    constexpr auto datatypes = cb::make_array(Datatype::Raw, Datatype::Xattr);

Is equivalent to:

    constexpr std::array<Datatype, 2> datatypes =
            {Datatype::Raw, Datatype::Xattr)};

Change-Id: Iff24b1037c611a0ca7573c11c98b38b72ab84a6a
Reviewed-on: http://review.couchbase.org/79838
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Will Gardner <willg@rdner.io>
Reviewed-by: Oliver Downard <oliver.downard@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>