Refactor: don't use cb_realloc in pipe
cb_realloc will free the old memory if it has to reallocate
the underlying buffer, causing a double free when the unique_ptr
tries to release it. I guess I must have been too tired when I
wrote the code to miss that.

Use new[] to allocate an array, and also "optimise" the realloc
functionality to pack the buffer before trying to reallocate it.

Change-Id: I25fbe8bd3c856a5bd3249986a9394a7007d6d98d
Reviewed-on: http://review.couchbase.org/81007
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>