MB-25442: Don't pack the buffer for every ensureCapacity
Packing the buffer means moving the data in the input buffer
to the beginning of the buffer.

The memcached core would call ensureCapacity first to make sure
that the header is there, then one more time to make sure that
the actual body fits. Depending on where in the buffer we are
we could end up packing the buffer even if there is enough space
available already.

Change-Id: Iaa8d18555e98e42695bec58051d602de2bc35270
Reviewed-on: http://review.couchbase.org/81516
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>