Add wdata() and made produced and consumed public
In order to track down the performance regression moving to
this interface try to get rid of the potential overhead from
calling through the lambda methods.

The valgrind integration was removed as it wouldn't work
if the caller called wdata() and copied in data before
calling produce() as it would touch memory marked as locked.

Change-Id: I47799297c794f278cda79fd6dcb526189d19bc26
Reviewed-on: http://review.couchbase.org/81900
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Will Gardner <willg@rdner.io>