Register threads with phosphor on Windows
There is currently the potential in phosphor to trigger a segfault
if a thread has logged an event with phosphor, that thread shuts-down
and then tracing stops (as the phosphor tracelog will attempt to
dereference a pointer into that thread's TLS).

Through luck alone this won't be triggered with the current set of
events on Windows since no thread that logs an event will be shut
down before tracing is stopped for the last time. (The bucket
creation/deletion threads do not currently log any events and
all the memcached/ep worker threads are shutdown *after* tracing
stops.

Change-Id: Ia5bce591fbfed98d30d103a4ea19169fcecee852
Reviewed-on: http://review.couchbase.org/82634
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>