CachelinePadded: Move operator<< into cb:: namespace
To allow argument-dependent lookup to work correctly for
CachelinePadded's operator<< it needs to be in the same namespace as
one of the arguments - i.e. cb::

Move it to the cb namespace.

Note: This was necessary as during some refactoring in kv_engine we
encountered a situation where the operator couldn't be found.

Change-Id: I9558ef6592f5dff9559169178034a42a23308eb7
Reviewed-on: http://review.couchbase.org/82692
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>