MB-26047:[7-b] Refactor histogram.h to work with <chrono> durations
Part of a cross repo change.

Change MicrosecondHistogram to use uint64_t as Rep, to ensure the same
behaviour as provided by hrtime_t which was of the same type. Add new
alias for custom microseconds used.

Introduce cb::duration_limits, which aims to behave similarly to
std::numeric_limits, but for chrono durations.

Add a second template argument to the Histogram class(and to other
related classes and methods), which expects a template class to be
passed(std::numeric_limits for primitive types or cb::duration_limits
for durations).

Change-Id: Ic061e11a79b09663b980e173dc4191d46d5aa35d
Reviewed-on: http://review.couchbase.org/84744
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>