Allow allocator to be specified for compression
couchstore still use cb_malloc/cb_free in order to allocate
blocks, and it would be beneficial to use the same library
methods to deal with compression everywhere in our system
(as an extra bonus it means that once we've added support
for a new algorithm we may use it everywhere).

Previously cb::compression allocated all blocks by using
new char[xx], and the buffer kept a std::unique_ptr<char[]>.
That means that couchstore would have to allocate another
buffer and copy the data over, which isn't very optimal.

This patch modifies the Buffer so that it internally it use
either new[]/delete[] or cb_malloc/cb_free for buffer
allocation.

Change-Id: I43d5de8168cd4bbd2c8a3c1713455287d72e41b4
Reviewed-on: http://review.couchbase.org/86374
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>