Replace INCLUDE_DIRECTORIES with TARGET_INCLUDE_DIRECTORIES
It is preferred to use target_include_directories() instead of
include_directories() to ensure that header dependancies are explicit
(and we don't accidently depend on headers). See 'Effective Modern
CMake[1]' for background.

Replace include_directories() with target_include_directories() where
necessary, or simply remove include_directories() where it is
unnecessary - for example where a target already depends on a library
and the library has already declared its public includes.

To assist with this, an explicit 'platform_headers' interface library
has been created, targets which only wish to use platform's headers
(for example tests) can link to that library to get the correct search
paths.

[1]: https://gist.github.com/mbinna/c61dbb39bca0e4fb7d1f73b0d66a4fd1

Change-Id: I4483fc3d5c888ef7bf1ece2e875df5c90655d1a6
Reviewed-on: http://review.couchbase.org/90742
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>