UBSan: Fix failing RTTI checks when using DSOs
UBSan's vptr checker needs to be able to lookup RTTI information
(type_info) for all calls via virtual functions. However, a number of
tests which load DSOs (via dlopen) currently fail when run under
UBSan.

For example:

    build-ubsan/kv_engine/engine_testapp -C 0 -E build-ubsan/kv_engine/ep.so -T build-ubsan/kv_engine/ep_testsuite.so -v -e dbname=./ep_testsuite.value_eviction.db

    runtime error: member call on address 0x7fffee10e460 which does not point to an object of type 'FileOpsInterface'
    0x7fffee10e460: note: object is of type 'StatsOps'
    00 00 00 00  a8 64 d8 ef ff 7f 00 00  a0 f9 06 ee ff 7f 00 00  c0 8f 83 f3 ff 7f 00 00  00 00 00 00
                 ^~~~~~~~~~~~~~~~~~~~~~~
                 vptr for 'StatsOps'
    #0 0x7ffff35bc970 in BufferedFileOps::constructor(couchstore_error_info_t*, FileOpsInterface*, buffered_file_ops_params) couchstore/src/iobuffer.cc:334

The reason for this is that cb_dlopen() calls dlopen with the
RTLD_LOCAL flag; this means that symbols in the opened DSO are not
made available to the calling binary, or any other loaded
DSOs. Specifically, in the case of ep_testsuite it loads ep.so via
cb_dlopen(); and this includes type info for a number of
storage-related classes such as StatsFileOps.

To solve this, replace RTLD_LOCAL with RTLD_GLOBAL; so all external
symbols in ep.so are accessible.

Change-Id: Ibcc0082998fad492b6f9871d4b3581bb837ef56a
Reviewed-on: http://review.couchbase.org/93760
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>