MB-30029: Add basic sized delete operator
Jemalloc 5.1 defines a c++14 sized delete operator which will be used in
preference to the standard delete operator (when deleting a complete type).
We must interpose our own to ensure we call the delete hook and can track
deallocations properly.

The operator defined in this patch just calls the delete hook and free,
ignoring the size parameter.

Array forms of the new and delete operator have been overloaded in version
5.1. This requires us to overload our version to call cb_malloc and
cb_free appropriately.

For some reason, under TSan, adding the sized delete operator changes
the runtime linking order which causes RocksDB to start calling our
overload of malloc_usable_size. This is fine, but there exists a bug
where any call to malloc_usable_size when using the system allocator
is infinitely recursive causing a stack overflow segfault. Fix this by
only overriding the function when we are NOT using the system allocator.

Change-Id: Iab8cfc36fbefc7f0216ec130426dd6d6cbfe3ddd
Reviewed-on: http://review.couchbase.org/99992
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>