MB-31688: Remove the use of [cb::]ProcessClock
It is defined as:

    using ProcessClock = std::chrono::steady_clock

The name is shorter, but it adds an extra lookup to figure out
what it is. It used to be different things when not all of the
compilers supported std::chrono::steady_clock, but now that it
is no longer the case it should be removed.

Change-Id: I23029181209df80cea1c357bdf3bc93de3f6c2dd
Reviewed-on: http://review.couchbase.org/100803
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>