MB-32016: Add 'unsanitized' variant of platform libs
If one or more sanitizers are enabled (AddressSanitizer,
ThreadSanitizer, ...) then define an additional variant of platform
libs which has the sanitizers /disabled/.  This allows targets which
require platform, but which are incompatible with the sanitizers
(e.g. Erlang NIFs) to still be built and run when we are building the
rest of the Server with sanitizer support.

Targets which suffer from this limitation can link against
'platform_unsanitized' instead of 'platform' to obtain a library which
doesn't use any of the sanitizers.

Change-Id: I0e979b690af5a0dcf87373020999f3f6a170430f
Reviewed-on: http://review.couchbase.org/103244
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>