MB-30040: Introduce platform_thread.h
Pulling in folly is a massive PITA because of how much stuff is in
the platform.h header file. To make this easier, split platform.h
into a number of logically grouped header files that can be included
instead.

Move the thread and locking declarations into their own header file.

Change-Id: I38f14d9e6c1653c31ac516f0618a909123ca4b72
Reviewed-on: http://review.couchbase.org/105526
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>