MB-30040: Introduce platform_socket.h
Pulling in folly is a massive PITA because of how much stuff is in
the platform.h header file. To make this easier, split platform.h
into a number of logically grouped header files that can be included
instead.

Move the socket declarations into their own header file. Don't move
them to "socket.h" as that's a C++ header file and there are C
consumers of these declarations.

Change-Id: I65163833f0265a98b1baee6bcbba1f3c0be71eda
Reviewed-on: http://review.couchbase.org/105543
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>