MB-23086: Add ArenaMalloc introspection functions
We need functions to retrieve statistics for:

1) An engine's "memory" statistics
2) An engine's "allocator" statistics
3) An engine's fragmentation

For 1) we can now obtain some detailed stats for a bucket. Additionally
we expose a way to ask for "non-bucket" memory stats.

For 2) we can now obtain verbose allocator dump for all arenas

For 3) we return a subset of 1) in size_t form so that the defragger
can log these details.

Change-Id: I4f4f1d6f171890dcd71683a67b6e8c9bf064c56a
Reviewed-on: http://review.couchbase.org/119220
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>