MB-33846: Make NonNegativeCounter deal with negatives
In the hash table in kv_engine we often add negative values.
This causes the stats to underflow. Correct this by doing similar
logic done in fetch_sub in fetch_add, and allowing fetch_add to
take a signed type. Also, allow fetch_sub to take a negative value
for consistency.

Change-Id: Ic63012ce41fe13a627477a386551aa7805ca00e2
Reviewed-on: http://review.couchbase.org/108202
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>