MB-23086: Create platform_cbassert library
Part of MB-23086 will be the enabling of a more efficient thread local
storage model known as 'initial-exec'. One inconvenience that
'initial-exec' introduces is that shared objects can no longer be safely
dynamically opened (e.g. dlopen/dlsym).

couchdb has in the past had issues which required the explicit
disablement of 'initial-exec' in jemalloc (MB-33042) so to prevent
further issues this commit takes couchdb's platform dependencies
(cbassert and backtrace) and places them in a  separate library.

Change-Id: I88a7d20134b4f2d572ce1d1c7161229c9b7e9ee6
Reviewed-on: http://review.couchbase.org/109564
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>