MB-23086: Change thread local storage to initial-exec
Compile platform using the initial-exec thread local storage model, a
more efficient form of thread local storage.

An inconvenience of this change (which doesn't impede our use of
platform) is that we lose the ability to safely dlopen the platform
shared library. Changes preceding this commit have removed any direct or
indirect dlopen of platform.

Change-Id: I56ad2fedf77885916aae20f5c8454c1ada17819c
Reviewed-on: http://review.couchbase.org/113572
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>