MB-23086: Allow per thread switch tcache disable
KV-engine has the need to disable the tcache when defragmenting, this
must be done on the defragmenter thread only without affecting
other thread running the engine or other engines.

To achieve this, add a new parameter to the switchToClient method
which allows a single switchToClient call to set the tcache status.

Change-Id: I56d6dec21d64a1509fc69fa0e3fef8e61290b1aa
Reviewed-on: http://review.couchbase.org/120135
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>