MB-23086: Reduce calls to tls_get_addr
With platform built as a shared library and dlopen safe, TLS
require a function call to retrieve a pointer to each
thread_local.

This call is generally considered 'expensive' and this patch
aims to reduce the number of calls made to get_tls_addr.

Examining the generated code for je_arena_malloc and some observations.

When thread_local is used on a non POD type, at least two calls to
get_tls_addr are needed.
  - A guard variable is checked (possible for construct/deconstruct)
  - The actual usage of the variable
  - Note the two calls maybe because of:
    * https://gcc.gnu.org/bugzilla/show_bug.cgi?id=82803
    * It looks like the guard and usage could all be done with 1 call

Firstly changing the type of ThreadLocalData from a std::unique_ptr
reduces the number of get_tls_addr calls in many paths.

Secondly by taking a copy of the currentClient in our
malloc/calloc/free wrappers (combined with the type change for tld)
reduced the number of calls on that path to a single get_tls_addr call.

Additional reduction of get_tls_addr calls was made in switchToClient
by changing the type of tcacheIds from std::array to a plain array. Now
the switchToClient fast path only encounters two get_tls_addr calls
on gcc 7.4, it seems this should be 1 call (and much experimenting with
taking pointers to the thread local address were tried), but suspect
that https://gcc.gnu.org/bugzilla/show_bug.cgi?id=82803 could be forcing
the two calls, one for access to tcacheIds and the second for the write
to currentClient.

Change-Id: Ia02ee91aedc14304e4a1958abaa55a4da2199dbd
Reviewed-on: http://review.couchbase.org/120483
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>