MB-37484: Make cb_malloc a weak symbol on Linux/OSX
Make cb_malloc (and friends) a weak symbol when built
for UNIX targets.

The outcome of this change is that any users of cb_malloc
that only link against libplatform_so.so (via CMake's platform
library target) will get a cb_malloc that uses libc malloc (and
means all operator new/delete replacements also end up in libc
malloc)

This fixes the Python crash because when libcouchstore.so is
dlopened there can be no way to get a mismatch of je_malloc
and malloc via our operator new/delete replacement.

For targets that do want the memory tracking allocator as the
implementation of operator new/delete and cb_malloc they will
be required to explicitly link against a new archive,
platform_cb_malloc.

platform_cb_malloc is a small library that also implements
cb_malloc and directs calls into cb::ArenaMalloc. The weak
symbol in platform_so.so gets "ignored" and the target will
now have all memory allocation directed through the
cb::ArenaMalloc (and for most builds that means je_malloc).

Windows builds don't use this technique, that is because the
weak symbol directive isn't as straight forward as on
OSX/Linux. So on Windows, when python dynamically loads
libcouchstore.dll (and the dependency libplatform.dll) operator
new/delete is bound to cb:ArenaMalloc. This because we do not
believe that this MB applies to Windows because of the way each
library/executable will 100% resolve to the correct operator
new/delete (this is our best assumption based on Windows
documentation and linker discussions).

Change-Id: Ic300ba2a8d89d7d25622e01190fa644bf5995200
Reviewed-on: http://review.couchbase.org/121140
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>