MB-37096: Add setupWindowsDebugCRTAssertHandling()
Add a new function to platform - setupWindowsDebugCRTAssertHandling().

If running on Windows with a Debug build, direct error and assertion
messages from the CRT to stderr, in addition to the default GUI
dialog box. Also prints a backtrace to the error.

Ensures that errors from Debug-mode tests etc are visible even if
running in a non-graphical mode (e.g. Jenkins CV job).

No-op on non-Windows, or if not a Debug build.

Change-Id: Ic71e918f8af4ccf6b5dc8d687eb447f43ec75567
Reviewed-on: http://review.couchbase.org/121349
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>